refactor(compiler): combine compiler codes into one place
BREAKING CHANGE
Now both `isolatedModules: true` and `isolatedModule: false` are in one single place regarding to codebase.
